refactor: 🔥 move parsing code into separate library dfraw_json_parser
