+++ src-tauri/Cargo.toml
@@ -15,14 +15,7 @@ rust-version = "1.57"
 tauri-build = { version = "1.0", features = [] }
 
 [dependencies]
-serde_json = "1.0"
-serde = { version = "1.0", features = ["derive"] }
 tauri = { version = "1.0", features = ["dialog-open", "fs-read-dir", "updater", "window-set-title"] }
-encoding_rs = "0.8"
-encoding_rs_io = "0.1"
-regex = "1.5"
-slug = "0.1"
-walkdir = "2"
 log = "0.4.17"
 
 [dependencies.simple_logger]
@@ -30,18 +23,12 @@ default-features = false
 version = "2.2.0"
 features = ["colors"]
 
-[dependencies.lazy_static]
-git = "https://github.com/rust-lang-nursery/lazy-static.rs.git"
-rev = "e83d66490775eb953807c39d1f073a6daa3bc760"
+[dependencies.dfraw_json_parser]
+git = "https://github.com/nwesterhausen/dfraw_json_parser.git"
 
 [dependencies.tauri-plugin-store]
 git = "https://github.com/tauri-apps/tauri-plugin-store"
 #branch = "main"
-
-[dependencies.phf]
-version = "0.10"
-default-features = true
-features = ["macros"]
 
 [features]
 # by default Tauri runs in production mode
