--- src-tauri/src/parser/raws/biomes.rs
@@ -1,86 +1,0 @@
-use phf::phf_map;
-
-pub static BIOMES: phf::Map<&'static str, &'static str> = phf_map! {
-    "MOUNTAIN" => "Mountain",
-"MOUNTAINS" => "Mountain",
-"GLACIER" => "Glacier",
-"TUNDRA" => "Tundra",
-"SWAMP_TEMPERATE_FRESHWATER" => "Temperate Freshwater Swamp",
-"SWAMP_TEMPERATE_SALTWATER" => "Temperate Saltwater Swamp",
-"MARSH_TEMPERATE_FRESHWATER" => "Temperate Freshwater Marsh",
-"MARSH_TEMPERATE_SALTWATER" => "Temperate Saltwater Marsh",
-"SWAMP_TROPICAL_FRESHWATER" => "Tropical Freshwater Swamp",
-"SWAMP_TROPICAL_SALTWATER" => "Tropical Saltwater Swamp",
-"SWAMP_MANGROVE" => "Mangrove Swamp",
-"MARSH_TROPICAL_FRESHWATER" => "Tropical Freshwater Marsh",
-"MARSH_TROPICAL_SALTWATER" => "Tropical Saltwater Marsh",
-"FOREST_TAIGA" => "Taiga",
-"TAIGA" => "Taiga",
-"FOREST_TEMPERATE_CONIFER" => "Temperate Coniferous Forest",
-"FOREST_TEMPERATE_BROADLEAF" => "Temperate Broadleaf Forest",
-"FOREST_TROPICAL_CONIFER" => "Tropical Coniferous Forest",
-"FOREST_TROPICAL_DRY_BROADLEAF" => "Tropical Dry Broadleaf Forest",
-"FOREST_TROPICAL_MOIST_BROADLEAF" => "Tropical Moist Broadleaf Forest",
-"GRASSLAND_TEMPERATE" => "Temperate Grassland",
-"SAVANNA_TEMPERATE" => "Temperate Savanna",
-"SHRUBLAND_TEMPERATE" => "Temperate Shrubland",
-"GRASSLAND_TROPICAL" => "Tropical Grassland",
-"SAVANNA_TROPICAL" => "Tropical Savanna",
-"SHRUBLAND_TROPICAL" => "Tropical Shrubland",
-"DESERT_BADLAND" => "Badlands",
-"DESERT_ROCK" => "Rocky Wasteland",
-"DESERT_SAND" => "Sand Desert",
-"OCEAN_TROPICAL" => "Tropical Ocean",
-"OCEAN_TEMPERATE" => "Temperate Ocean",
-"OCEAN_ARCTIC" => "Arctic Ocean",
-"POOL_TEMPERATE_FRESHWATER" => "Temperate Freshwater Pool",
-"POOL_TEMPERATE_BRACKISHWATER" => "Temperate Brackish Pool",
-"POOL_TEMPERATE_SALTWATER" => "Temperate Saltwater Pool",
-"POOL_TROPICAL_FRESHWATER" => "Tropical Freshwater Pool",
-"POOL_TROPICAL_BRACKISHWATER" => "Tropical Brackish Pool",
-"POOL_TROPICAL_SALTWATER" => "Tropical Saltwater Pool",
-"LAKE_TEMPERATE_FRESHWATER" => "Temperate Freshwater Lake",
-"LAKE_TEMPERATE_BRACKISHWATER" => "Temperate Brackish Lake",
-"LAKE_TEMPERATE_SALTWATER" => "Temperate Saltwater Lake",
-"LAKE_TROPICAL_FRESHWATER" => "Tropical Freshwater Lake",
-"LAKE_TROPICAL_BRACKISHWATER" => "Tropical Brackish Lake",
-"LAKE_TROPICAL_SALTWATER" => "Tropical Saltwater Lake",
-"RIVER_TEMPERATE_FRESHWATER" => "Temperate Freshwater River",
-"RIVER_TEMPERATE_BRACKISHWATER" => "Temperate Brackish River",
-"RIVER_TEMPERATE_SALTWATER" => "Temperate Saltwater River",
-"RIVER_TROPICAL_FRESHWATER" => "Tropical Freshwater River",
-"RIVER_TROPICAL_BRACKISHWATER" => "Tropical Brackish River",
-"RIVER_TROPICAL_SALTWATER" => "Tropical Saltwater River",
-"SUBTERRANEAN_WATER" => "Underground caverns (in water)",
-"SUBTERRANEAN_CHASM" => "Underground caverns (out of water)",
-"SUBTERRANEAN_LAVA" => "Magma sea",
-"ALL_MAIN" => "All biomes excluding pools, rivers, and underground features",
-"ANY_LAND" => "All main biomes excluding oceans and lakes",
-"ANY_OCEAN" => "All ocean biomes",
-"ANY_LAKE" => "All lake biomes",
-"ANY_TEMPERATE_LAKE" => "All temperate lake biomes",
-"ANY_TROPICAL_LAKE" => "All tropical lake biomes",
-"ANY_RIVER" => "All river biomes",
-"ANY_TEMPERATE_RIVER" => "All temperate river biomes",
-"ANY_TROPICAL_RIVER" => "All tropical river biomes",
-"ANY_POOL" => "All pool biomes",
-"NOT_FREEZING" => "All land biomes excluding Mountain, Glacier, and Tundra",
-"ANY_TEMPERATE" => "All Temperate land biomes - marshes, swamps, forests, grassland, savanna, and shrubland",
-"ANY_TROPICAL" => "All Tropical land biomes - marshes, swamps (including Mangrove), forests, grassland, savanna, and shrubland",
-"ANY_FOREST" => "All Forest biomes (excluding Taiga)",
-"ANY_SHRUBLAND" => "Temperate and Tropical Shrubland",
-"ANY_GRASSLAND" => "Temperate and Tropical Grassland",
-"ANY_SAVANNA" => "Temperate and Tropical Savanna",
-"ANY_TEMPERATE_FOREST" => "Temperate Coniferous and Broadleaf Forests",
-"ANY_TROPICAL_FOREST" => "Tropical Coniferous and Dry/Moist Broadleaf Forests",
-"ANY_TEMPERATE_BROADLEAF" => "Temperate Broadleaf Forest, Grassland/Savanna/Shrubland, Swamps, and Marshes",
-"ANY_TROPICAL_BROADLEAF" => "Tropical Dry/Moist Broadleaf Forest, Grassland/Savanna/Shrubland, Swamps (including Mangrove), and Marshes",
-"ANY_WETLAND" => "All swamps and marshes",
-"ANY_TEMPERATE_WETLAND" => "All temperate swamps and marshes",
-"ANY_TROPICAL_WETLAND" => "All tropical swamps and marshes",
-"ANY_TROPICAL_MARSH" => "All tropical marshes",
-"ANY_TEMPERATE_MARSH" => "All temperate marshes",
-"ANY_TROPICAL_SWAMP" => "All tropical swamps (including Mangrove)",
-"ANY_TEMPERATE_SWAMP" => "All temperate swamps",
-"ANY_DESERT" => "Badlands, Rocky Wasteland, and Sand Desert",
-};
