--- src-tauri/src/parser/raws/mod.rs
@@ -1,4 +1,0 @@
-pub mod biomes;
-pub mod creature;
-pub mod names;
-pub mod tags;
